IsNotConnected into own component, cleanup
